調整route切換轉導相對應元件
