+++ resume/src/images/carousel/contentPasta.png
