+++ resume/src/images/carousel/pasta.jpeg
