+++ resume/src/views/HomePage/index.jsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 
 import Header from '../../components/Header/Header';
 import Footer from '../../components/Footer/Footer';
-import HomeContent from '../../components/Content/HomeConponents/HomeContent';
 
 /**
  * 個人首頁
